more thing

+++ include/pl-unistd.h
@@ -71,8 +71,8 @@ extern long pl_syscall(int syscall, int paramNum, void* params);
 
 int open(const char* pathname, int flags, int mode);
 int close(int fd);
-int read(int fd, void* buffer, unsigned int length);
-int write(int fd, void* data, unsigned int length);
+int read(int fd, void* buffer, unsigned long length);
+int write(int fd, void* data, unsigned long length);
 void exit(int code);
 void* mmap(void* addr, unsigned long length, int prot, int flags, int fd, long offset);
 int munmap(void* addr, unsigned long length);
